EFL build fix after r163729.
* Modules/indexeddb/IDBDatabase.cpp:
(WebCore::IDBDatabase::version):


Canonical link: https://commits.webkit.org/146504@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163736 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/Modules/indexeddb/IDBDatabase.cpp
@@ -137,7 +137,7 @@ uint64_t IDBDatabase::version() const
 {
     // NoIntVersion is a special value for internal use only and shouldn't be exposed to script.
     // DefaultIntVersion should be exposed instead.
-    return m_metadata.version != IDBDatabaseMetadata::NoIntVersion ? m_metadata.version : IDBDatabaseMetadata::DefaultIntVersion;
+    return m_metadata.version != IDBDatabaseMetadata::NoIntVersion ? m_metadata.version : static_cast<uint64_t>(IDBDatabaseMetadata::DefaultIntVersion);
 }
 
 PassRefPtr<IDBObjectStore> IDBDatabase::createObjectStore(const String& name, const Dictionary& options, ExceptionCode& ec)
